Point light now exist

+++ 3Dproj/plHelper.h
@@ -19,4 +19,4 @@ bool loadHShader(std::string name, ID3D11Device* device, ID3D11HullShader*& hSha
 bool loadDShader(std::string name, ID3D11Device* device, ID3D11DomainShader*& dShader);
 bool CreateTexture(std::string file, ID3D11Device* device, ID3D11Texture2D*& tex, ID3D11ShaderResourceView*& texSRV);
 bool CreateInputLayoutOwn(ID3D11Device* device, ID3D11InputLayout*& inputLayout, std::string& VbyteCode, D3D11_INPUT_ELEMENT_DESC inputDesc[], int nrOfElements);
-bool CreateTextureCube(std::string files[], ID3D11Device* device, ID3D11Texture2D*& tex, ID3D11ShaderResourceView*& texSRV, float wh);
+bool CreateTextureCube(std::string files[], ID3D11Device* device, ID3D11Texture2D*& tex, ID3D11ShaderResourceView*& texSRV, UINT wh);
